More refactoring, moved SearchInstance into separate files. Code cleanup.
