+++ src/ui/include/Search/advancedsearchdock.h
@@ -5,13 +5,11 @@
 #include <QDockWidget>
 #include <QAbstractButton>
 #include <QString>
-#include <QTreeWidget>
 
 #include <memory>
 #include <vector>
-#include <map>
 
-#include "filesearcher.h"
+#include "searchinstance.h"
 
 class QWidget;
 class QLayout;
@@ -21,55 +19,6 @@ class QComboBox;
 class QLineEdit;
 class QCheckBox;
 
-class SearchInstance : public QObject {
-    Q_OBJECT
-
-public:
-    SearchInstance(const SearchConfig& config);
-    ~SearchInstance();
-
-    // Getters
-    bool getShowFullLines() const { return m_showFullLines; }
-    bool areResultsExpanded() const { return m_resultsAreExpanded; }
-    bool isSearchInProgress() const { return m_isSearchInProgress; }
-
-    QTreeWidget*        getResultTreeWidget() const { return m_treeWidget.get(); }
-    const SearchConfig& getSearchConfig() const { return m_searchConfig; }
-    const SearchResult& getSearchResult() const { return m_searchResult; }
-
-    SearchResult getFilteredSearchResult() const;
-
-    // Actions
-    void expandAllResults();
-    void collapseAllResults();
-
-    void selectNextResult();
-    void selectPreviousResult();
-
-    void showFullLines(bool showFullLines);
-    void copySelectedLinesToClipboard() const;
-
-signals:
-    void searchCompleted();
-    void resultItemClicked(const DocResult& doc, const MatchResult& result);
-
-private:
-    void onSearchProgress(int processed, int total);
-    void onSearchCompleted();
-
-    bool m_isSearchInProgress = true; // Search is started in the constructor so it can default to true
-    bool m_resultsAreExpanded = false;
-    bool m_showFullLines = false;
-
-    SearchConfig    m_searchConfig;
-    std::unique_ptr<QTreeWidget> m_treeWidget; // TODO: use Qt's parent system instead
-    SearchResult    m_searchResult;
-    FileSearcher*   m_fileSearcher = nullptr;
-
-    // These map each QTreeWidget item to their respective MatchResult or DocResult
-    std::map<QTreeWidgetItem*, const MatchResult*>  m_resultMap;
-    std::map<QTreeWidgetItem*, const DocResult*>    m_docMap;
-};
 
 
 /**
