+++ src/ui/ui.pro
@@ -102,7 +102,8 @@ SOURCES += main.cpp\
     nqqrun.cpp \
     Search/filesearcher.cpp \
     Search/filereplacer.cpp \
-    Search/searchobjects.cpp
+    Search/searchobjects.cpp \
+    Search/searchinstance.cpp
 
 HEADERS  += include/mainwindow.h \
     include/topeditorcontainer.h \
@@ -146,7 +147,8 @@ HEADERS  += include/mainwindow.h \
     include/nqqrun.h \
     include/Search/filesearcher.h \
     include/Search/searchobjects.h \
-    include/Search/filereplacer.h
+    include/Search/filereplacer.h \
+    include/Search/searchinstance.h
 
 FORMS    += mainwindow.ui \
     frmabout.ui \
